% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashpcaR-package.R
\docType{package}
\name{flashpcaR-package}
\alias{flashpcaR-package}
\alias{flashpcaR}
\title{flashpcaR
An R interface to FlashPCA for fast principal component analysis and
related analyses.}
\description{
\tabular{ll}{
Package: \tab flashpcaR\cr
Type: \tab Package\cr
Version: \tab 2.0.1\cr
Date: \tab 2017-01-19\cr
License: \tab GPL (>= 3)\cr
}
}
\references{
Abraham G, Inouye M (2014) Fast Principal Component Analysis of
Large-Scale Genome-Wide Data. PLoS ONE 9(4): e93766.
doi:10.1371/journal.pone.0093766
}
\author{
Gad Abraham and Yixuan Qiu
Maintainer: Gad Abraham <gad.abraham@unimelb.edu.au>
}
