% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca.R
\name{print.cv.scca}
\alias{print.cv.scca}
\title{Prints a cv.scca object}
\usage{
\method{print}{cv.scca}(x, ...)
}
\arguments{
\item{x}{A cv.scca object to be printed}

\item{...}{Ignored}
}
\description{
Prints a cv.scca object
}
